typescript intital commit
