+++ package.json
@@ -2,16 +2,21 @@
   "name": "snipsync",
   "version": "0.0.2-beta-6",
   "description": "Sync docs with github repo code snippets",
-  "main": "index.js",
+  "main": "index.ts",
   "repository": "git@github.com:temporalio/snippetsync.git",
   "author": "Cully Wakelin <cully@temporal.io>",
   "license": "MIT",
   "private": false,
   "scripts": {
-    "test": "echo \"tests are not available\" && exit 1"
+    "test": "echo \"tests are not available\" && exit 1",
+    "build": "tsc",
+    "dev": "ts-node index.ts",
+    "start": "nodemon ./dist/index.js",
+    "prod": "yarn build && yarn start"
   },
   "dependencies": {
     "@octokit/rest": "^18.0.3",
+    "@octokit/types": "^6.1.1",
     "anzip": "^0.2.0",
     "arraybuffer-to-buffer": "^0.0.7",
     "cli-progress": "^3.8.2",
@@ -23,6 +28,12 @@
     "rimraf": "^3.0.2"
   },
   "bin": {
-    "snipsync": "./src/index.js"
+    "snipsync": "index.ts"
+  },
+  "devDependencies": {
+    "@types/cli-progress": "^3.8.0",
+    "@types/node": "^14.14.16",
+    "ts-node": "^9.1.1",
+    "typescript": "^4.1.3"
   }
 }
