--- src/File.js
@@ -1,9 +1,0 @@
-class File {
-  constructor(filename) {
-    this.filename = filename;
-    this.lines = [];
-  }
-}
-
-module.exports = { File };
-
