--- src/common.js
@@ -1,12 +1,0 @@
-module.exports = {
-  rootDir: process.cwd(),
-  cfgFile: 'snipsync.config.yaml',
-  extractionDir: 'sync_repos',
-  markdownCodeTicks: '```',
-  fmtStartCodeBlock: (ext) => '```' + ext,
-  readStart: '@@@SNIPSTART',
-  readEnd: '@@@SNIPEND',
-  writeStart: '<!--SNIPSTART',
-  writeEnd: '<!--SNIPEND',
-  fmtProgressBar: (message) => `⭐ + | {bar} | {percentage}% | {value}/{total} chunks | ${message}`
-}
