+++ src/config.ts
@@ -1,9 +1,9 @@
-const { join } = require('path');
-const { sync } = require('node-read-yaml');
-const { cfgFile, rootDir, fmtProgressBar } = require('./common');
-const progress = require ('cli-progress');
+import { join } from 'path';
+import { sync } from 'node-read-yaml';
+import { cfgFile, rootDir, fmtProgressBar } from './common';
+import progress from 'cli-progress';
 
-module.exports.readConfig = () => {
+export const readConfig = () => {
   const cfgPath = join(rootDir, cfgFile);
   const cfgProgress = new progress.Bar({
     format: fmtProgressBar(`loading configuration from ${cfgPath}`),
